stoock delete and pagination

+++ inventory-genius-backend/src/main/java/com/example/inventorygenius/service/StockService.java
@@ -2,21 +2,30 @@ package com.example.inventorygenius.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.example.inventorygenius.entity.Item;
 import com.example.inventorygenius.entity.Stock;
+import com.example.inventorygenius.entity.StockCount;
 import com.example.inventorygenius.entity.Storage;
 import com.example.inventorygenius.repository.StockRepository;
+import com.example.inventorygenius.service.StockCountService;
+
+
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Optional;
 
 @Service
 public class StockService {
 
     @Autowired
     private StockRepository stockRepository;
+
+    @Autowired
+    private StockCountService stockCountService;
 
     // Method to add a new item
     public Stock addStock(Stock stock) {
@@ -28,9 +37,28 @@ public class StockService {
         return stockRepository.findAll();
     }
 
+    @Transactional
     public void deleteStockById(Long id) {
-        stockRepository.deleteById(id);
+        Optional<Stock> stockOptional = stockRepository.findById(id);
+        
+        if (stockOptional.isPresent()) {
+            Stock stock = stockOptional.get();
+            StockCount sc = stockCountService.getStockCountBySKUCode(stock.getSkucode());
+            Double prevCount = sc.getCount();
+    
+            if (Double.parseDouble(stock.getAddQty()) > 0) {
+                sc.setCount(prevCount - Double.parseDouble(stock.getAddQty()));
+            } else if (Double.parseDouble(stock.getSubQty()) > 0) {
+                sc.setCount(prevCount + Double.parseDouble(stock.getSubQty()));
+            }
+    
+            stockCountService.updateStockCount(sc);
+            stockRepository.deleteById(id);
+        } else {
+            throw new NoSuchElementException("Stock with ID " + id + " not found");
+        }
     }
+    
 
     public Stock updateStock(Long id, Stock stockDetails) {
         Stock stock = stockRepository.findById(id)
